refactored code and added test cases
