+++ kairon/api/app/routers/bot/data.py
@@ -10,7 +10,6 @@ from kairon.events.definitions.content_importer import DocContentImporterEvent
 from kairon.events.definitions.faq_importer import FaqDataImporterEvent
 from kairon.exceptions import AppException
 from kairon.shared.auth import Authentication
-from kairon.shared.channels.whatsapp.bsp.dialog360 import BSP360Dialog
 from kairon.shared.cognition.data_objects import CognitionSchema
 from kairon.shared.cognition.processor import CognitionDataProcessor
 from kairon.shared.concurrency.actors.factory import ActorFactory
@@ -369,46 +368,6 @@ async def upload_doc_content(
     if is_event_data:
         event.enqueue()
     return {"message": "Document content upload in progress! Check logs."}
-
-
-
-@router.post("/upload/media_upload", response_model = Response)
-async def upload_media_file_content(
-    file_content: UploadFile = File(...),
-    current_user: User = Security(Authentication.get_current_user_and_bot, scopes = DESIGNER_ACCESS),
-):
-    """
-    Handles the upload of file content for processing, validation, and eventual storage.
-    """
-    MongoProcessor.validate_media_file_type(file_content)
-    error_message, file_path = await MongoProcessor.media_handler_save_and_validate(
-        bot = current_user.get_bot(),
-        user = current_user.get_user(),
-        file_content = file_content,
-    )
-
-    if error_message:
-        return Response(success=False, message="Validation failed", data=error_message, error_code=400)
-
-    media_id = await MongoProcessor.upload_media_to_bsp(
-        bot = current_user.get_bot(),
-        user = current_user.get_user(),
-        file_path = file_path,
-        file_info = file_content,
-    )
-
-    return Response(message = "File uploaded successfully!", data = media_id)
-
-@router.get("/upload/media_upload", response_model=Response)
-async def get_media_ids(
-    current_user: User = Security(Authentication.get_current_user_and_bot, scopes=DESIGNER_ACCESS),
-):
-
-    try:
-        media_ids = BSP360Dialog.get_media_ids(current_user.get_bot())
-        return Response(message="List of media ids", data=media_ids)
-    except Exception as e:
-        return Response(success=False, message=f"Failed to fetch media ids: {e}", data=None, error_code=500)
 
 @router.get("/content/error-report/{event_id}", response_model=Response)
 async def download_error_csv(
