+++ kairon/shared/data/constant.py
@@ -397,24 +397,24 @@ LOG_TYPE_METADATA = {
 }
 
 MIME_TYPE_LIMITS = {
-    "audio/aac": "16 MB",
-    "audio/amr": "16 MB",
-    "audio/mpeg": "16 MB",
-    "audio/mp4": "16 MB",
-    "audio/ogg": "16 MB",
-    "text/plain": "100 MB",
-    "application/vnd.ms-excel": "100 MB",
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": "100 MB",
-    "application/msword": "100 MB",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document": "100 MB",
-    "application/vnd.ms-powerpoint": "100 MB",
-    "application/vnd.openxmlformats-officedocument.presentationml.presentation": "100 MB",
-    "application/pdf": "100 MB",
-    "image/jpeg": "5 MB",
-    "image/png": "5 MB",
-    "image/webp": "500 KB",
-    "video/3gpp": "16 MB",
-    "video/mp4": "16 MB",
+    "audio/aac": 16 * 1024 * 1024,
+    "audio/amr": 16 * 1024 * 1024,
+    "audio/mpeg": 16 * 1024 * 1024,
+    "audio/mp4": 16 * 1024 * 1024,
+    "audio/ogg": 16 * 1024 * 1024,
+    "text/plain": 100 * 1024 * 1024,
+    "application/vnd.ms-excel": 100 * 1024 * 1024,
+    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": 100 * 1024 * 1024,
+    "application/msword": 100 * 1024 * 1024,
+    "application/vnd.openxmlformats-officedocument.wordprocessingml.document": 100 * 1024 * 1024,
+    "application/vnd.ms-powerpoint": 100 * 1024 * 1024,
+    "application/vnd.openxmlformats-officedocument.presentationml.presentation": 100 * 1024 * 1024,
+    "application/pdf": 100 * 1024 * 1024,
+    "image/jpeg": 5 * 1024 * 1024,
+    "image/png": 5 * 1024 * 1024,
+    "image/webp": 500 * 1024,
+    "video/3gpp": 16 * 1024 * 1024,
+    "video/mp4": 16 * 1024 * 1024,
 }
 
 class LogType(str, Enum):
