+++ kairon/shared/data/processor.py
@@ -3,7 +3,7 @@ import itertools
 import re
 import shutil
 
-import aiofiles
+
 import ujson as json
 import os
 import uuid
@@ -172,13 +172,11 @@ from ..callback.data_objects import CallbackConfig, CallbackLog, CallbackRespons
 from ..chat.broadcast.data_objects import MessageBroadcastLogs
 from ..cognition.data_objects import CognitionSchema, CognitionData, ColumnMetadata
 from ..constants import KaironSystemSlots, PluginTypes, EventClass, EXCLUDED_INTENTS, \
-    WhatsappBSPTypes
 from ..content_importer.content_processor import ContentImporterLogProcessor
 from ..custom_widgets.data_objects import CustomWidgets
 from ..importer.data_objects import ValidationLogs
 from ..live_agent.live_agent import LiveAgentHandler
 from ..log_system.base import BaseLogHandler
-from ..log_system.factory import LogHandlerFactory
 from ..multilingual.data_objects import BotReplicationLogs
 from ..test.data_objects import ModelTestingLogs
 
@@ -8959,84 +8957,6 @@ class MongoProcessor:
                 error_message['Extra columns'] = f"{extra_columns}."
 
         return error_message
-
-    @staticmethod
-    async def media_handler_save_and_validate( bot: Text, user: Text, file_content: File):
-        """
-        Saves the media file and validates its type.
-
-        :param bot: The bot ID
-        :param user: The user ID
-        :param file_content: The uploaded file
-        :return: A dictionary of error messages if validation fails
-        """
-        error_message = {}
-
-        content_dir = os.path.join("media_upload_records", bot)
-        Utility.make_dirs(content_dir)
-        file_path = os.path.join(content_dir, file_content.filename)
-
-        async with aiofiles.open(file_path, "wb") as buffer:
-            while chunk := await file_content.read(1024 * 1024):
-                await buffer.write(chunk)
-
-        await file_content.seek(0)
-
-        return error_message, file_path
-
-    @staticmethod
-    async def upload_media_to_bsp(bot: str, user: str, file_path: str, file_info: File):
-        """
-        Uploads the file to BSP and deletes the temporary local file.
-        """
-        from ..channels.whatsapp.bsp.factory import BusinessServiceProviderFactory
-
-        media_id = None
-        try:
-            bsp_type= WhatsappBSPTypes.bsp_360dialog.value
-            media_id = await BusinessServiceProviderFactory.get_instance(bsp_type).upload_media_file(bot, bsp_type, user,
-                                                                file_info.filename, file_info.content_type,
-                                                                file_info.size)
-            logger.info(f"Media uploaded successfully: {media_id}")
-        except Exception as e:
-            logger.error(f"Error uploading file to BSP: {str(e)}")
-            raise AppException(f"Media upload failed: {str(e)}")
-
-        finally:
-            if file_path:
-                try:
-                    Utility.remove_file_path(file_path)
-                except Exception as cleanup_err:
-                    logger.warning(f"Failed to cleanup temp file {file_path}: {cleanup_err}")
-
-        return media_id
-
-    @staticmethod
-    def validate_media_file_type(file_content):
-        content_type = file_content.content_type
-
-        if content_type not in MIME_TYPE_LIMITS:
-            raise AppException(
-                f"Invalid file type: {content_type}. "
-                f"Allowed types are: {', '.join(MIME_TYPE_LIMITS.keys())}."
-            )
-
-        size_limit_str = MIME_TYPE_LIMITS[content_type]
-        size_limit, unit = size_limit_str.split()
-        size_limit = int(size_limit) * {
-            "KB": 1024,
-            "MB": 1024 * 1024,
-        }[unit]
-
-        file_content.file.seek(0, 2)
-        size = file_content.file.tell()
-        file_content.file.seek(0)
-
-        if size > size_limit:
-            raise AppException(
-                f"File size {size / (1024 * 1024):.2f} MB exceeds the "
-                f"limit of {size_limit / (1024 * 1024):.2f} MB for {content_type}."
-            )
 
     def get_column_datatype_dict(self, bot, table_name):
         from ..cognition.processor import CognitionDataProcessor
